+++ docs/source/changelog.rst
@@ -9,6 +9,7 @@ Changelog
 Version (dev)
 -------------
 
+* Make it clear that :class:`srctools.fgd.UnknownHelper` should not be relied on.
 * Treat ``ogg`` files as valid sounds in various places.
 * Consistently raise :class:`TypeError` if invalid types are sent to a :class:`~srctools.tokenizer.Tokenizer`.
 * Properly add ``base_plat_train`` class resource function and fix the ``item_teamflag``.